requirements.txt
tensorflow==2.15.0
scikit-learn==1.4.2
fastapi==0.110.0
uvicorn==0.29.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.6.4
requests==2.31.0
python-multipart==0.0.9
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
opencv-python-headless>=4.9.0
matplotlib>=3.8.0
tqdm>=4.66.0
streamlit>=1.33.0
